MobileDetection/requirements.txt
requests==2.22.0
Keras==2.2.5
detecto==1.1.2
Flask==1.1.1
opencv_python_headless==4.1.2.30
pandas==1.0.3
numpy==1.17.3
tensorflow==1.13.1
Pillow==7.1.1
